Refactor classes into different files
