+++ Assignment-3/Assignment3Server/Category.cs
@@ -1,0 +1,12 @@
+﻿using System.Text.Json.Serialization;
+
+namespace categoryNS
+{
+    public class Category
+    {
+        [JsonPropertyName("cid")]
+        public int Id { get; set; }
+        [JsonPropertyName("name")]
+        public string Name { get; set; }
+    }
+}
