+++ Assignment-3/Assignment3Server/Request.cs
@@ -1,0 +1,10 @@
+﻿namespace requestNS
+{
+    public class Request
+    {
+        public string Method { get; set; }
+        public string Path { get; set; }
+        public string Date { get; set; }
+        public string Body { get; set; }
+    }
+}
