+++ Assignment-3/Assignment3Server/Response.cs
@@ -1,0 +1,8 @@
+﻿namespace responseNS
+{
+    public class Response
+    {
+        public string Status { get; set; }
+        public string Body { get; set; }
+    }
+}
